monitor/backend/requirements.txt
# S001-Pro Monitor Backend
# FastAPI + SQLite + JWT

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-dotenv==1.0.0
httpx==0.25.2

# 开发依赖
pytest==7.4.3
pytest-asyncio==0.21.1
